+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix writing batch v2, [PR-1033](https://github.com/reductstore/reductstore/pull/1033)
+
 ## 1.17.3 - 2025-11-17
 
 ### Added